create a locale for svm
locale for svm was added and it was proven it is
a subset of the lipschitz locale. Some needed lemmas for the
hinge loss were proven, like convexity and lipschitzness.
The proof for the lipschitzness does not followe the book,
so the lemmas for the subgradients are not needed for the svm theory.